Rework get_vector
The interface is now more consistent, querying for a single variable
now behaves the same way as does querying one attribute with a single
valid value, i.e., the value column is not renamed to the variable name.
Also, the columns are omitted by default but can be enabled when needed.